consolidate buttons into the same component, dynamic text for apply button
