+++ components/index.js
@@ -1,6 +1,5 @@
-export { ApplyLink } from './apply-link'
 export { default as ArticlePreview } from './article-preview'
-export { default as Button } from './button'
+export { Button, ApplyLink } from './button'
 export { default as CompanyLogo } from './company-logo'
 export { default as Footer } from './footer'
 export { default as Form } from './form'
